Move constants to recommendations package
