+++ src/main/java/com/autotune/analyzer/recommendations/engine/KruizeRecommendationEngine.java
@@ -17,6 +17,7 @@ package com.autotune.analyzer.recommendations.engine;
 
 import com.autotune.analyzer.kruizeObject.RecommendationSettings;
 import com.autotune.analyzer.recommendations.Recommendation;
+import com.autotune.analyzer.recommendations.RecommendationConstants;
 import com.autotune.analyzer.utils.AnalyzerConstants;
 import com.autotune.common.data.result.ContainerData;
 
@@ -26,7 +27,7 @@ import java.util.HashMap;
 public interface KruizeRecommendationEngine {
     public String getEngineName();
     public String getEngineKey();
-    public AnalyzerConstants.RecommendationCategory getEngineCategory();
+    public RecommendationConstants.RecommendationCategory getEngineCategory();
     public HashMap<String, Recommendation> generateRecommendation(ContainerData containerData, Timestamp monitoringEndTime, RecommendationSettings recommendationSettings);
     public void validateRecommendations();
     public boolean checkIfMinDataAvailable(ContainerData containerData);
